pyproject.toml
[tool.black]
line-length = 79
target-version = ["py37", "py38"]

[tool.poetry]
name = "invoicex"
version = "1.0.0"
description = ""
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.7.1"
Jinja2 = "^3.1.1"
requests = "^2.23"
python-dotenv = "^0.20.0"
pandas = "^1.4.2"
pandoc = "^2.1"
tabulate = "^0.8.9"
websockets = "^8.1"
gql = "^3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
